pyproject.toml
[project]
name = "xeofs"
version = "2.4.1"
description = "Comprehensive EOF analysis in Python with xarray: A versatile, multidimensional, and scalable tool for advanced climate data analysis"
authors = [
    {name = "Niclas Rieger", email = "niclasrieger@gmail.com"},
    {name = "Samuel J. Levang", email = "slevang@gmail.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "numpy>=1.24",
  "pandas>=2",
  "xarray>=2023.04.0",
  "scikit-learn>=1.0.2",
  "tqdm>=4.64.0",
  "dask>=2023.0.1",
  "statsmodels>=0.14.0",
  "netCDF4>=1.5.8",
  "numba>=0.57",
  "typing-extensions>=4.8.0",
  "zarr>=2.14.0",
  "xarray-datatree>=0.0.12",
]

[project.optional-dependencies]
dev = [
  "build>=1.0.0",
  "ruff>=0.3",
  "pytest>=7",
  "pytest-xdist>=3",
  "coverage>=6",
  "pre-commit>=3",
  "nbstripout>=0.6",
  "pooch>=1.6.0",
]
docs = [
  "rpy2>=3.5",
  "sphinx-gallery>=0.14",
  "sphinx-design>=0.6",
  "sphinx-copybutton>=0.5",
  "nbsphinx>=0.9",
  "pydata-sphinx-theme>=0.15",
  "sphinx>=8",
  "nbconvert>=7.9",
  "myst-parser>=3.0",
  "matplotlib>=3.4",
  "seaborn>=0.11",
  "cartopy>=0.22",
  "pooch>=1.6.0",
  "ipython>=8.14",
  "ipykernel>=6.23",
]

[project.urls]
homepage = "https://github.com/xarray-contrib/xeofs"
repository = "https://github.com/xarray-contrib/xeofs"
documentation = "https://xeofs.readthedocs.io/en/latest/"

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42",
  "setuptools-scm>=7",
]

[tool.ruff]
target-version = "py311"

[tool.semantic_release]
version_toml = [ "pyproject.toml:project.version" ]
version_source = "tag"
commit_message = "chore(release): v{version}"
build_command = """
    python -m pip install build~=0.10.0
    python -m build .
"""
branch = "main"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
